modularize the layout template
this commit adds a number of jinja2 {% block %} elements to the main
layout template, so that individual pages can extend or override this
content.  The use case I am planning for is adding custom per-page
javascript for handling client-side filtering of tables.  With these
blocks in place, a page can add additional scripts like this:

    {% block scripts %}
    {{ super() }}
    <script src=".../myscript.js"></script>
    {% endblock %}