fail with exit code when static generation fails
Previously, failure to generate static output would not result in a
nonzero exit code, so our tests would not detect failures of the 'ara
generate' command.

This commit allows any exceptions generated by freeze() to percolate
up to cliff, which will present an error message and exit with a
nonzero code.

This commit also contains a few minor cleanups to the generate
command:

- Renames 'ara/static.py' to 'ara/freezer.py', because we also have a
  'static' directory.

- If path is required, make it a positional parameter rather than an
  optional parameter.

- Emit messages using logging rather than 'print'.

Closes: #86