Only try to load facts if there are any in the first place
This resolves an issue where a host with no facts (setup module
or gather_facts never run throughout the entire playbook) would
generate 404's -- we don't want to generate a 404, rather just
not display the facts.

This is stopgap bugfix to allow static generation to work when
there are hosts without facts. The UI/UX around facts is planned
to be improved later already.

Change-Id: I381ad346d7bc9b80eb8aecd4f1f459ff31e93f67
Story: 2000764