Implement ara_record data types
Data types allows a user to specify a type when recording data.
Depending on the type, the display behavior or format will be different
in the web interface.
The first implemented types are as follows:
- text (default): straight text in <pre>
- url: value becomes a hyperlink
- json: value is json pretty-printed

Other example types that could be implemented eventually would be
markdown or rst for rich formatting.

Change-Id: I28b78a2b5899ece3b0abc4648bd8d0a2678c80ee