Move unit and integration tests in-tree
This is mostly no-op, there are minor fixes because the tests
directory was not included in pep8 and bandit checks.
We'll now be checking tests with flake8.

Tests were already included in the source tarball. When installing
from pypi, this means tests will now be shipped as well.
In binary packaging that is being working on, there will be separate
packages for ara, docs and test. This makes this part slightly more
straightforward.

Change-Id: Ia8b7f8fc2a815d5e974d597b0d0d3c2b7a66220d