Stop requiring pymysql to be installed by default
By adding pymysql to our requirements, we're including what is
essentially an optional backend and we're being opinionated about
what users should be using.

In reality, we're using sqlalchemy so users should be able to
consume sqlite (default) or things like mysql or pymysql.

Let's keep some level of documentation about pymysql for now
(backwards-compat-ish) and iterate on this further eventually.

Change-Id: I14476b5f1e2498a7451d61e940816baa2b2bc89f