Get rid of Flask-Testing dependency
This commit rewrites tests to get rid of our dependency on flask-testing.
flask-testing is in a weird spot in terms of packaging, it depends on
an old (last released 2014) python dependency: twill.

flask-testing doesn't really abstract all that much and isn't overly
useful so let's rewrite things to get it out of the way.
This is required as part of the binary packaging that is in process
for ARA.

Change-Id: Ia35c181d3669e6a11a270de0aabeda56d43af2a9