Default undefined values to None when templating CLI fields with Jinja
When we got a null/None values for a field in a CLI command, we would
default it to a Jinja Undefined object.
In pretty tables, this worked out okay: the value was just empty.

However, when sending data that has strict(er) types such as JSON
or YAML, this would cause errors like "Undefined is not JSON serializable".
This is a problem when requesting output in different formats through
cliff.

On top of fixing this behavior, this commit also adds some integration
tests to ensure that different formatting outputs will work as expected.

Change-Id: I255e9e5a96788baa73e61a0e15996275f033ade7