Properly default the database and logs location to ARA's directory
Both the database and the log location configuration both defaulted
properly to ~/.ara. However, if the location of the ARA directory
was changed, their default would remain under ~/.ara.

Change this to make it so the default locations for the database
and logs properly follow if the location of the ARA directory is
changed so that we don't need to update all three settings.

Change-Id: I715224d68d08d8af7cf1487d3c41fa78ac75550a