Support generating html or junit for one or many specific playbooks
This provides a --playbook parameter to the ara generate html
and ara generate junit commands in order to generate a static
report only for one or many specified playbooks.

This is especially useful in the context of a persistent or central
environment where a user might store a lot of playbook run results.

Change-Id: I42feeaa9829a44e1add690c774041e3aff17eeea