Defer data tables rendering with AJAX pre-processed results
It turns out that rendering thousands of items on a single page
was awful due to both the server and the client needing to render
things.
Part of this problem had already been addressed by offloading the
large content (task result, host facts, files, etc.) into separate
pages with iframes. However, we still rendered data tables with
potentially thousands of rows -- and even more so if the user has
more than a playbook per page.

We're fixing this behavior by pre-processing the different tables'
data and having the data tables fetch them asynchronously over AJAX.
This operation is non-blocking and the page will show up immediately
with loading spinning wheels if the content is not ready.

This also adds some better handling when there are no resources for
a given type (no hosts, no plays, etc.).

Change-Id: Ic16b6f75cbec2a627eb95417a6e9d3509058a143