Fix display bug in hosts tab
Enclosing the tooltip in double quotes caused the double quotes inside
the tooltip to stop rendering early. Use single quotes instead.

Additionally, now that result rendering is deferred and ran
asynchronously, we need to use the datatables "fnInitComplete"
callback to ensure that tooltips are generated properly once the
datatable is ready.

Change-Id: Iccfd0431cbc577b3a0573f5d261294ef27494aa5