Cleanup ARA configuration module
Ansible 2.3 is coming soon and the get_config method we rely on to
retrieve parameters is changing. We already had limited support for
it but now it's cleaner and more reliable.

Otherwise, this is just cleaning up what has been organically growing
to be a bit messy over time. As part of that cleaning up, we're also
removing dead code around the jinja filter to truncate path lengths.

Change-Id: I16301382d67531a57613cac2e42887e07431acf4