Filter empty generation warnings by default
This commit adds a new config 'ignore_empty_generation'.
This will default to true and will filter MissingURLGeneratorWarning
warnings raised from flask-frozen.

The reason why we default to true is because the generation does not
fail -- it can, however, have one or multiple endpoints with no data.
For example, ARA users which do not leverage the ara_record module
would see a warning because there would not be any data for the
recorded data endpoint.

If generation would occur from an empty database, it would
essentially just generate the home page that says the app could not
find any records -- which is the same behavior as the database-driven
web app, and that's fine. In the end, the warning was mostly just
confusing users into thinking something failed.

Change-Id: I71ebad1ac5c3f3f8052940dc616d35dac29c4a37