Use Ansible's CallbackBase._dump_results to dump results
The _dump_results method has some internal logic which we want to
use. The problem this is meant to address is that it happens to be
the burden of the callbacks to respect the "no_log" directive.

The _dump_results method takes care of removing "internal" Ansible
keys as well as strip data if no_log is used, amongst other things.

Change-Id: Ic914b73fdf3df13c4f258f4dbf8d5675f98cacb1