Improve count performance
It turns out that the sqlalchemy .count() query method isn't
particularly fast and this appears to be a known fact.
Use an optimized way of retrieving the amount of records
instead.
This isn't ground breaking but it's still significant when
viewing a large amount of data -- any amount of performance
increase is definitely welcome.

On a fairly large dataset:
- Average home before: 0.031s
- Average home after: 0.019s

- Average reports before: 7.208s
- Average reports after: 5.279s

Change-Id: Ib9bb3a264c38c2c383e9a88a531743c1ad1bf682