Append slash to modal iframe URLs
This fixes a weird issue that was noticed in some circumstances when
the browser would try to load HTTP iframes in what would otherwise be
served under HTTPS.

Let's take for example an iframe location which would be "/file/uuid".
For the purposes of compatibility between the dynamic web application
and the statically generated version, this path is actually exposed
as a directory.

The web application handles this properly, it's a known route.
However, in the case of the statically generated version, it's ambiguous
and it's up to the webserver to handle it.
Apache, for example, realizes the requested path is a directory and will
issue a HTTP 301 redirect to the right destination properly.
However, nginx doesn't do this automatic redirection and since the actual
path is "/file/uuid/index.html", it fails to load.

Change-Id: I7b14fa919e5544543153db8566dde75c2f82ce80
Story: 2000947