Hide foreground INFO alembic.runtime.migration stdout logging
The main purpose was to send message such as:
- Context impl SQLiteImpl
- Will assume non-transactional DDL

To the ARA logfile and hide them from stdout, same thing for the
SQL migrations.
This particular patch gets us something sort of in the middle,
it hides the two messages above from stdout but does not send them
to the logfile.

The SQL migrations are also printed to stdout, not sent to the
log file.
We need to revisit this later and do a better job at logging in
general.

Change-Id: I22d28e590228b61e525fe154c2138918dcf64e3b