Perf: Only count things once and do not iterate playbook list twice
This is a continuation in spirit of
https://review.openstack.org/#/c/450449/ .

Before this patch, we'd count things more than once and this had a
performance impact, especially when dealing with a large amount of
data.
We'd also iterate across the list of playbooks twice, now we only
do it once.

This is now fixed.

Change-Id: Icbfb31ef4ea1267207ecc9081dbffb8f8f335ef5