Implement framework for packaged js/css libs, use packaged jquery
This commits adds a route for /static/packaged/<module>/<file>
which provides the requested file through the XStatic module.
The list of available modules is provided by the app configuration.

This removes the need for bundling the jquery js file in-tree as
we're now relying on the packaged xstatic version of it.

Change-Id: I1a0a813f06bb62343b17fe9116d82ec614e01fac