Add permanent links for playbook reports
This adds a permanent link on the left hand side of each report
in the report list allowing users to link to a specific report.

The templates were re-arranged to avoid duplication.

The static generation performance is not noticeably affected by
this change but it does yield a slightly heavier tree.
For a static report of 35 playbooks with 5000 results and 82 hosts:
- Before: 54MB (26MB gzipped)
- After: 67MB (30MB gzipped)

The tradeoff in terms of feature are significant enough to consider
the increase in weight a non-issue.

Change-Id: I05b2ec3c38cb265426ba1198e2d5f0b5b54b6b20