Replace file table by a proper file hierarchy tree
It's apparently insanely complicated to recurse across a list of
paths and generate a JSON structure suitable for the
bootstrap-treeview javascript library in order to generate the
file hierarchy involved in the playbooks.

This takes an approach with pyfakefs where we essentially fake/mock
out the directories and files in memory and then properly traverse
them with "native" os.path functions.

This patch bundles bootstrap-treeview js and css, this will be
taken out as soon as it is packaged as XStatic on PyPi.

Change-Id: I24b948bf488f26f91d5fea4b0e75fe8ca8b9afd9