Properly store task results provided from loops
This adds support for saving task results for tasks in which a loop
(ex: with_items) is used. Before that, ARA would only save the
global result of the task and a minimal message returned by Ansible.

Note that there is a known issue in which ARA might miss the last
item of a task, this is an issue reported upstream [1].

[1]: https://github.com/ansible/ansible/issues/24207

Change-Id: I23565ea5d81731ccdef4337d8a925b7ef88d1b5a