Add playbook parameters and metadata, handle task tags
We now retrieve all the CLI options used when running
ansible-playbook, if they are available.
Users can prevent CLI options from being saved by ARA
with the new configuration option, ARA_IGNORE_PARAMETERS.

By default, we do not save extra_vars as it is likely to
contain sensitive information such as passwords.

These parameters can be found in a new parameter panel
for each report and is effective only for new playbook
runs. Previous playbook runs cannot be retro-fitted.

Additionnally, we now store task tags.
Task tags are shown when hovering over a task name that
has a task associated with it and when viewing the
task result in the popup modal.
It is also possible to search for a tag in the task
result search box.

The UI is a bit rough around the edges, it'll be
improved in a later commit.

Change-Id: Icb7f175efbf87a411ad998ea18172748fae44035