Default the home page to reports, move home to "about"
When you're browsing ARA, you're really interested in the reports.
The home page was a nice way to explain at a quick glance what was
ARA and the data it made available to the users but it was sort of
annoying for experienced users.
It still fits an important purpose so we're moving that page to
"/about", leaving it unchanged.

The playbook reports are now the default page.

Change-Id: I5a21a71edb2a006eb2b6efc1a51a4af2216cd378