Streamline integration tests
We've been doing things we don't really need to be doing.. installing
and uninstalling packages, creating files, etc.
The built-in integration tests is meant to be a quick smoke test to
tell that ARA is not horribly broken.

The ARA code review now gates against openstack-ansible which
certainly provides much better coverage in the context of using
Ansible in different ways.

Change-Id: Ibd4e7ba67c2067ea3dca311f6eefeb075af41be9