Use the escape filter instead of the safe filter where appropriate
The safe filter was used to escape potentially problematic characters
so that content could be browsed without problems.
However, it was found that the safe filter could allow user-generated
content to be interpreted by the browser such as a javascript script
from a task result content.

Replace this by the escape filter which suits the purpose by replacing
characters with HTML encoded equivalents without potentially interpreting
content.

Remaining usages of the safe filter are appropriate, for example in the
CLI, we don't want to replace characters with HTML equivalents.

Change-Id: I268863f5cd5c243d01cb3f486a115b587727ed75
Co-Authored-By: Tristan Cacqueray <tdecacqu@redhat.com>
Co-Authored-By: David Moreau Simard <dmsimard@redhat.com>