Revert using the escape filter instead of safe for files
In eb4df93db003eae5daff66c649338d6e72943970, we replaced the usage
of the safe filter to the escape filter in order to prevent html or
javascript code from being interpreted (XSS).

Pygments' HTMLFormatter already takes care of escaping things so
by using the escape filter, we're double escaping and this is causing
the *actual* pygments html code to be escaped.

Change-Id: Ic0abf0e24f3d28e2c4a6b5a85f8f07cc3d2ddc91