Py3: Fix bad Parent/Child one to many relation in the model
We have many tasks for one file and one file for many tasks.
Fix the backwards relationship.

It is unknown why it worked in py2, but it suddenly started
failing in python3 with the error:
"Instance File is already associated with an instance of Task via
its Task.file attribute, and is only allowed a single parent."

Co-Authored-By: Lars Kellogg-Stedman <lars@redhat.com>
Co-Authored-By: David Moreau Simard <dmsimard@redhat.com>
Change-Id: I4822833d6fcc6a6d1d74e27e98f7b980d16ab53a