Rename the notion of task_results to results
This is a bit of a nitpick which will hopefully end up making things
cleaner overall.
We don't have different type of results, we only have task results.
As such, it's not necessary to mention that these are task results, it's
implicit.

This also changes the primary key of the results table from a randomly
generated uuid to an auto-increment integer which will make the database
faster to scan and search for.

Change-Id: Ifa5623fc82108922acb069c1bad46cfb4e44a19d