Re-license ARA from ASLv2 to GPLv3
ARA was originally licensed as Apache 2.0 software to align with
the license used by the larger community of OpenStack software.

However, ARA is fairly coupled with Ansible, which is GPLv3.
By maintaining two licenses simultaneously, we may unknowingly infringe
on the incompatibility of the licenses by importing or linking code
to and from Ansible software.

In order to avoid the headaches involved in maintaining two licenses
side by side, remove the ambiguity as to what is the effective license,
we are relicensing ARA to be fully GPLv3.

Since ARA is not an official OpenStack project, it does not have an
impact on it's status within the OpenStack community [1] and will
remain as a hosted project.

We are doing this license change now rather than later as doing this
change later becomes increasingly complicated as the project gains more
code contributors.

[1]: http://lists.openstack.org/pipermail/legal-discuss/2017-July/000474.html

Change-Id: I2ec1e0a1d1b38fc74dcb6af2b4f81226d1faa0df