Ignore 404's when generating static report with frozen-flask
With the API, there are now multiple situations where it is possible
that a 404 is returned, for example during generation of an
incomplete playbook report.
A 404 is therefore not fatal and should not interrupt the generation.

Change-Id: I4e28fb4f8055027b98201023d218ffe3c3a1e39d
(cherry picked from commit 3b5db8b9311ce2efa7dc98c66f7ef6b2783f881e)