Include reference to childs instead of actual childs for playbook API
The playbook API now returns references (id, link) to it's children
resources rather than the children resources themselves.
This is more lightweight since including every children resource would
be very heavy.

Squashed together is another significant change to move the "default"
endpoint for each resource from "/api/v1/<resource>" to
"/api/v1/<resource>/" to keep parity with static generation because
flask-frozen needs to know whether to create a file or a directory if
there are children resources. By suffixing a '/', we are making it a
directory instead of a file.

Change-Id: Icd806bfb7a590bd591e986ed2e109b3218b3f5e4