Get rid of the HostFacts database table
Just like the Stats database table that we removed [1], the HostFacts
database table ran with the assumption that hosts were not unique per
playbook as was the case in the early ARA prototyping.

Now that hosts are unique per playbook, it no longer makes sense to
have a separate table to track facts per host across playbook runs.

[1]: https://review.openstack.org/#/c/491563/

Change-Id: I4f8a29f3156f82a16a88872c11c46124e63b2158