Validate, retrieve and set relationships instead of foreign keys
For example, instead of setting playbook_id, we'll retrieve the
actual Playbook model, validate it exists and set the playbook
relationship instead of playbook_id.

Setting playbook_id isn't equivalent, it just sets the foreign key
and it misses out on potential relationships as defined in the model
like one to many, many to one, etc.

Where possible, we're now also recovering parents from existing
relations -- for example, we don't need a playbook_id for a result
because we can recover the playbook model from the task model.
It also doesn't ever make sense to be able to patch the playbook_id
from an item so take that out.

Change-Id: I19cbae58e1883e6c93dc50762d5ffea3e11b05f8