Show names of tasks when showing results
When running "ara result list" show also task names field,
like it's done with "ara task list".
It's needed when trying to parse output to understand which
tasks failed and which didn't, according to their names in play.
As private case it's needed to stats collection, when you want
to send stats of passed tasks only to graphite.

Change-Id: I01fb06b80514445c05f589b026cb98637c80f748