Allow configuring logging via config file
The fileConfig stanza sets up alembic migration logging with the in-tree
logging config file. In some cases (zuul) that winds up stepping on
other logging config.

Add a config setting that can be used to pass a logging config file.
This file can be in logging ini format, or in yaml or json format. If
given it will apply settings to ara and alembic migrations.

If a file is not given, the current logging behavior remains. However,
logging config was removed from alembic.ini and instead set during
configure_logging so that the migrations don't step on other logging
config.

Change-Id: Ie555ab6ba6a177d084e62cd1bf2b89954894fea1