Automatically tag playbook files, create or get file if it exists
We don't need to handle the logic of tagging playbook files at the
callback level. We can automatically tag them if the path of the file
matches the path of the playbook.

Additionally, ensure that posting a file that already exists simply
returns the existing file, emulating a "get_or_create" call.

Change-Id: I1a05780d6c0a2c097339a3ddf9b8ef0916aa039e