Properly decode non-ascii content and re-serialize results as required
We weren't properly decoding non-ascii content, so that's now fixed.
Results can be provided either as dicts (one result) or as a list of
dicts (multiple results: i.e, with_items). Serialize the results
parameter as json which will take care of creating a dict or a list
of dicts as appropriate.

Change-Id: Ie1a4140f64aa63a94c6248ee5b2e99decfb1dd17