Introduce a client with backend selection
This is squashed in order to make my life less miserable.

This introduces a formal python client which is able to communicate
with the ARA API either internally (flask's internal client) or
the ARA HTTP/REST API.
The client takes care of handling the differences between the internal
and the REST API and offers a single and unique interface.
Sending and retrieving data through the client, no matter if using the
internal or the REST API, works exactly the same.

The client will use the internal interface by default but can be
made to use the http interface with the ARA_API_CLIENT configuration
parameter.

The REST API endpoint defaults to the embedded webserver URL and can
be customized with the ARA_API_ENDPOINT configuration parameter.

Change-Id: Idd00003b85a66a2560cdfb7243a11cbf6afc4ecd