Delete IncludeRole object from result object for include_role tasks
The IncludeRole object crashes the callback because ARA tries to
serialize the result._result dict and result._result['include_role']
is not serializable, it's an IncludeRole class instance.

We don't need this key and what is inside so let's just work around
the issue by deleting it for now.

See: https://github.com/ansible/ansible/issues/30385
Change-Id: I5b00d45244f7c90aa4a70171db88e19c94be5a9b