Ansible 2.4: Properly load the configuration file parser
Due to a misunderstanding of how configuration loading occured in
Ansible 2.4, we did not load the configuration file. It turns out
we have to do it ourselves now so let's do that.

This also requires a fix in the Ansible get_config method which
was merged [1] and backported for 2.4.1.0.
We're not going to work around the bug in Ansible 2.4.0.0 and as such,
blacklist that version and coordinate a new ARA release with Ansible
2.4.1.0.

[1]: https://github.com/ansible/ansible/pull/31200

Change-Id: I691ef992c9c6b698f03927e2b69de26f0c1d57d1