Add a sqlite middleware to load logged sqlite databases on-the-fly
The use case for this middleware is especially interesting for large
scale users that generate and host large static reports, for example
in continuous integration.

To put this use case into perspective, it was "benchmarked" against
a single job from the OpenStack-Ansible project:

- 4 playbooks
- 4647 tasks
- 4760 results
- 53 hosts, of which 39 had host facts
- 416 saved files

This data is contained in a single and unique 5.6MB (2.9MB gzipped)
ARA sqlite database.

Generating a static report from that database takes ~1min30s on an
average machine. It weighs 63MB (27MB recursively gzipped), contains
5321 files and 5243 directories.

This middleware allows a user to store only the sqlite database
and provide what ends up being the same end result.

Change-Id: I3558c50769f735e64b7ee10aad7f38b17fe2d5a6