Re-structure API routes and responses to match a list/detail model
This adds new routes for the API and changes the returned responses
to be a list/detail model.

For example, when querying the playbooks endpoint, we would get a list
of playbooks and the entirety of their details in one call.
This was very heavy because even one playbook can contain thousands of
hosts, tasks and results -- then you could have a large list of
playbooks each with those thousands of elements.

API calls that are meant to return list of items now contain the
minimum information and references to the items they contain.

For example:

/api/v1/playbooks -> Returns a list of playbook IDs and hrefs
/api/v1/playbooks/1 -> Returns details for playbook #1
/api/v1/playbooks/1/files -> Returns a list of file IDs and hrefs
                             for playbook #1

This makes the API responses much lighter, faster and easier to
generate.

Change-Id: I5580c268df53c4be2f6f7f21fcf103655719376b