Import ARA configuration, views and API "just-in-time"
Doing otherwise broke some assumptions because, for example,
"ara.config" loads the configuration automatically on import.

This is also preferable because these don't need to be loaded if we
have a flask current_app context, so only load them if they need to be.

This should also have a positive impact on the performance.

Change-Id: I43bac3b6af6830c0fb7c3ee6b0408de31ec72388