Refactor configuration and drop static generation
This is a large commit but one that would have been fairly complicated
to split.

Removal of the static HTML generation feature
---------------------------------------------
We realized that while the feature was quite useful, it really didn't
scale very well [1] and was also accountable for a fair amount of
hacks in order to make it work.
Wanting to keep 100% feature parity between the dynamic and static
versions of the reports, this meant holding back some features that
would not have worked in the static reports.

The use case of being able to host individual reports easily is a very
important one and that is why we will put emphasis on making the
"sqlite over http" backend work as simple as possible.

Configuration refactor
----------------------
- Several configuration parameters have been removed or otherwise
  "hardcoded" to a dynamic default where appropriate.
- Configuration parameters have been split in three categories to
  start working the split of ARA components: "base", "logging" and
  "webapp".
- ARA now ships a default (and vastly improved!) logging configuration
- Flask context_processors, filters and errorhandlers have been folded
  back into the webapp module

Cleanup
-------
- Remove dependencies for Decorator and Debtcollector, turns out we
  haven't been using those for a while
- Cleaned up hacks for running flask-frozen such as trailing slashes,
  "fake" html/txt mime types and index for hosts, results and files.

[1]: https://ara.readthedocs.io/en/latest/advanced.html

Change-Id: Id82141c4cd4e4d8b98f493756c6716d861251389