Instanciate the configuration loading into classes
It turns out that, either by design or for some other reason, Flask's
app.config.from_object only ever imports the object once, even if
a second instance of the application is instanciated.

This is a problem because the configuration of the applicatoin can
change between two instances of an application.
For example, in the same unit test session, we might want to test
different database backends.

By instanciating the configuration, we can ensure they are loaded only
when the instances are created.

Change-Id: I94d0266123a3000612502181582531b461717b56