Fix POST /playbooks/:id/files
The Playbook is kind of special in the sense that it has it's own
(playbook) file but it's also very convenient to store references for
all the files relevant to a playbook execution in a single place.

This makes it so we are able to append files individually as they come
up in the various stages of the execution of a playbook.

Change-Id: I7bf8dcb8286adb702d0a884cdcfeea6ac419b417