Disable database automatic creation/migration in the sqlite middleware
The sqlite middleware is meant to serve existing databases. It's not
expected that we need to create a new database and make sure sql
migrations are up to date.
Disabling this yielded on non-scientific benchmarks an improvement on
the order of 300ms to 500ms being cut for the initial app bootstrap.

Change-Id: Iedcf3e7ef5181b6c6aa4275a2cfb23146994fe5b