Add the concept of reports to the API
A report is a generic container meant to group or correlate different
playbooks.
It could contain a single playbook run or a group of playbook runs.
It can also be used to represent phases or dynamic tagging of playbook
runs.
For example, you could have a report named "failures" and make it so
failed playbooks are added to this report, for example.

The main purpose of this is to make reports generic and dynamic from
an API standpoint so we can build on top of this later.

Change-Id: I398f0337987abe31fa1e886f66ec9c3e644a32d6