Logging: don't disable existing loggers, avoid conflicting
An issue reported to upstream Ansible [1] highlighted that it was
fairly easy for the logging configurations of Ansible and ARA to
conflict with each other.

This commit makes it so we no longer set the root logger -- we don't
need to since we're not expecting orphaned loggers which would inherit
from the root configuration.

It also makes it so we don't disable existing loggers which would
prevent other software from being able to log properly.

And finally, it prefixes the names of the handlers and formatters
by "ara_" in order to prevent clashing with other software.

[1]: https://github.com/ansible/ansible/issues/41838

Change-Id: Iecc3324cc10a767b607d3055d6b93a12b9ae87ee