Improve self-healing of missing playbook file references
Another way the issue of a missing playbook file reference could
manifest itself was if the file had been created but is_playbook
would not have yet been set.

Since we already have a unique constraint on file paths per playbook.id,
we are able to recover from this as well.

Change-Id: I271a6c3733972ee74ed5e709516c168a1ee66f55