Add an ARA_IGNORE_FACTS variable and stop saving ansible_env by default
The ansible_env fact is no longer saved by default as env variables are
often used to carry passwords, tokens or otherwise privileged information
that might not be relevant or suitable to save in ARA.

This is done through a new ARA_IGNORE_FACTS configuration parameter
which allows users to define a list of facts that they might not be
comfortable saving.

This prevents facts configured in this list from being saved to host
facts and task results.

Change-Id: I69c110706198b9fb6dd0324e2157859fe01e5b34