Add concept of host alias for regrouping hosts
Ansible doesn't provide a way to uniquely identify hosts which would
allow ARA to track a single host across multiple playbook runs.

This patch adds a field for each host, 'alias'.
The plan is to leverage this field as a way to regroup the results
for what is believed to be the same host -- the same alias.

A host alias is arbitrary and generic. It can be the inventory
hostname or the result of a hashing algorithm like md5sum/sha1sum
of a combination of host facts.

Change-Id: I16b668f473a9f31686c7a921d7dcd31a5292febd