Use 'ara-clients' instead of __name__ when determining client version
This resolves an exception by pbr when Ansible attempts to load the
client:
"""
Versioning for this project requires either an sdist tarball,
or access to an upstream git repository. It's also possible that
there is a mismatch between the package name in setup.cfg and
the argument given to pbr.version.VersionInfo. Project name
ara.clients.http was given, but was not able to be found.
"""

Change-Id: Id56bf0909d17259313307d433a6104fe02e0215e