Implement support for setting host aliases
Hosts in Ansible are unique per playbook because there is no concept
of persistence between different playbook runs. A host could be named
"webserver" and be a completely different host in two playbook runs.

In order to make it possible to track a given host across different
playbook runs, it is now possible to set a host alias.
Host aliases are explicitely provided by the user as the "ara_host_alias"
hostvar which defaults to the inventory hostname if it is not set.

Change-Id: Iaa3b0ce57968b9e4977d25a559e21e67a37b1d62