First iteration of authentication and security docs
This is a first iteration on basic security and authentication
documentation. It documents the security settings users might want to
pay attention to and explains how to configure them.

We're bumping the version of sphinx in order to enable the
"autosectionlabel" extension so we don't need to create explicit
labels for sections.

Change-Id: I694eb1a2aa03193335f11cbda40ed6962357c428