Replace the ansible-integration job by the new integration jobs
ara-infra now provides a new and improved set of integration jobs
so we no longer need to have the ansible-integration job.

These new jobs test ara-server, ara-clients and ara-plugins against
a variety of operating systems and versions of Ansible.

It was convenient to be able to run these tests from ara-server by
itself, we'll re-introduce a wrapper so we can continue to do so
later.

Depends-On: https://review.openstack.org/#/c/630303
Change-Id: Iedf259e44af5165be08c5d6917e5af98b281202d