Rely on the API server for the get_or_create methods
Instead of "manually" verifying if a file or a host has already
been created for the playbook, always send it and rely on the
API server to do the validation.

This is hopefully a good tradeoff: while we're now always sending
data, we save one API call whenever we want to get or create files
and hosts.

Change-Id: I5b3d4ddef0ee5514fc56d27a2729d926e432b4de