Add support for API authentication in the client
This adds support for authentication to the offline and http clients
as well as the plumbing for the callback and the ara_record module to
support it.

This works but is missing a lot of tests. New issues have been created
to make sure we don't forget about adding tests.

Fixes: https://github.com/ansible-community/ara/issues/16
Change-Id: I62937dac1eb349baf6dea386dbd22b59d52319cd
Co-Authored-By: David Moreau-Simard <dmsimard@redhat.com>
Co-Authored-By: Florian Apolloner <florian@apolloner.eu>