Add playbook.arguments to the /playbooks list view
Making playbook arguments available in the list view is a compromise
which increases the size of the response but prevents a call to the
playbook detailed view if the only purpose was to retrieve the playbook
arguments.

For example, this would allow ara-web to provide more context in the
playbook list view (ex: if --check was enabled) without needing additional
API calls for each playbook.

Change-Id: I32c2abfb2770b240ac2ea98c14ce054b922a4dfc