UI: Set API link to object when viewing object
When at the index, link to the API index.
Otherwise, when viewing a playbook, link to the playbook detail API or
when viewing a result, link to the result detail API, etc.

The APIRoot view was renamed to APIIndex because it didn't have a name
and was conflicting with the built-in "api-root" view provided by DRF.

Change-Id: Ifa1434dd65c70b409e88aaf22e6b337f751f61a6