Make env setup allow use of other plugins
Without this patch, evaluating the output of the .env function
will override any other plugins defined in env variables.

This is a problem, as it might even break a run, assuming the
env var is pointing to, for example, a output display callback
plugin, for which the path would then be unavailable.

This fixes it by ensuring that, if any of the vars to define
is already defined, the print will print first (followed by :) and
then adding ara into the more prefered spot for loading.

Change-Id: Ife50033be09746d38b209ae37abb30995def84d2
(cherry picked from commit 4aadf1a127859ff1de13d8b775073f35c55a9603)