UI: Use DRF views and serializers instead of the API client
Using the API client to retrieve data from the API for the UI would
have been nice but it complicates things because then we don't have
easy access to the web-rendering facilities that django provides like
pagination.

Going through the DRF and serializers also means we don't need to stand
up an API server -- even if ephemeral through the offline client.

Change-Id: I9352f3dda93d4e7704a71fdaca3d043a2fcf65c3