Add scripts to build and run ARA API containers
This contributes scripts to build and run ARA API server container images.

Both use a Fedora base image but one installs ARA from source while
the the other installs ARA from fedora packages.

The images are set up to use the gunicorn application server and include
the necessary libraries for users that wish to use the postgresql or
mysql backends.

The scripts are tested with the included playbooks and zuul job.

Change-Id: If85210395dd3d93c80da83fd69b86eecfa0185ef